Renamed "utils" package to "datastructures"
Renamed utils both in main and test.
This renaming only affects graphs and tests.
More details:
This change only affected newly added classes by this branch.
The renaming did NOT change classes outside of this branch,
because other classes did not yet use the new classes which
are added by this branch.